requirements-extended.txt
# Extended Requirements for Real-time Speech Translator
# Install with: pip install -r requirements-extended.txt

# Core dependencies (from requirements.txt)
sounddevice>=0.4.6
numpy>=1.24.0
openai-whisper>=20230314
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Voice Activity Detection
webrtcvad>=2.0.10

# REST API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Export Utilities
python-docx>=0.8.11
openpyxl>=3.1.2

# Batch Processing
tqdm>=4.65.0
watchdog>=3.0.0

# Audio Processing
soundfile>=0.12.1
scipy>=1.10.0
librosa>=0.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
requests>=2.31.0
